warnings

+++ src/Userland/SampleCodeModule/processModule.c
@@ -9,14 +9,14 @@ unsigned long int createProcess(char* name, int (*entry)(int, char**), int argc,
 }
 
 void kill(unsigned long int pid) {
-  systemCall((uint64_t)KILL, pid, 0, 0, 0, 0);
+  systemCall((uint64_t)KILL, (uint64_t)pid, 0, 0, 0, 0);
 }
 
 void getPS(tProcessData*** psVec, int *size) {
-  systemCall((uint64_t)PS, psVec, size, 0, 0, 0);
+  systemCall((uint64_t)PS, (uint64_t)psVec, (uint64_t)size, 0, 0, 0);
 }
 
 
 void waitpid(unsigned long int pid) {
-  systemCall((uint64_t)WAITPID, pid, 0, 0, 0, 0);
+  systemCall((uint64_t)WAITPID, (uint64_t)pid, 0, 0, 0, 0);
 }
